Relocated yt command and parser
